Split Assets
